finished typescript conversion, started formatting sass variables, defined types for ember-responsive and ember--table
