+++ addon/templates/components/gavant-table.hbs
