+++ app/components/gavant-table.js
@@ -1,0 +1,1 @@
+export { default } from '@gavant/ember-table/components/gavant-table';
