--- app/components/table.js
@@ -1,1 +1,0 @@
-export { default } from '@gavant/ember-table/components/table';
