+++ app/styles/gavant-ember-table.scss
@@ -1,60 +1,3 @@
-// Bootstrap - Core variables and mixins
-@import 'node_modules/bootstrap/scss/functions';
-@import 'node_modules/bootstrap/scss/variables';
-@import 'node_modules/bootstrap/scss/mixins';
-@import 'node_modules/bootstrap/scss/root';
-@import 'node_modules/bootstrap/scss/print';
-
-// Bootstrap - Core CSS
-@import 'node_modules/bootstrap/scss/reboot';
-@import 'node_modules/bootstrap/scss/type';
-@import 'node_modules/bootstrap/scss/images';
-@import 'node_modules/bootstrap/scss/code';
-@import 'node_modules/bootstrap/scss/grid';
-@import 'node_modules/bootstrap/scss/tables';
-@import 'node_modules/bootstrap/scss/forms';
-@import 'node_modules/bootstrap/scss/buttons';
-
-// Bootstrap - Components
-@import 'node_modules/bootstrap/scss/transitions';
-@import 'node_modules/bootstrap/scss/dropdown';
-@import 'node_modules/bootstrap/scss/button-group';
-@import 'node_modules/bootstrap/scss/input-group';
-@import 'node_modules/bootstrap/scss/custom-forms';
-@import 'node_modules/bootstrap/scss/nav';
-@import 'node_modules/bootstrap/scss/navbar';
-@import 'node_modules/bootstrap/scss/card';
-@import 'node_modules/bootstrap/scss/breadcrumb';
-@import 'node_modules/bootstrap/scss/pagination';
-@import 'node_modules/bootstrap/scss/badge';
-@import 'node_modules/bootstrap/scss/jumbotron';
-@import 'node_modules/bootstrap/scss/alert';
-@import 'node_modules/bootstrap/scss/progress';
-@import 'node_modules/bootstrap/scss/media';
-@import 'node_modules/bootstrap/scss/list-group';
-@import 'node_modules/bootstrap/scss/close';
-
-// Bootstrap - Components w/ JavaScript
-@import 'node_modules/bootstrap/scss/modal';
-@import 'node_modules/bootstrap/scss/tooltip';
-@import 'node_modules/bootstrap/scss/popover';
-@import 'node_modules/bootstrap/scss/carousel';
-
-// Bootstrap - Utility classes
-@import 'node_modules/bootstrap/scss/utilities/align';
-@import 'node_modules/bootstrap/scss/utilities/background';
-@import 'node_modules/bootstrap/scss/utilities/borders';
-@import 'node_modules/bootstrap/scss/utilities/clearfix';
-@import 'node_modules/bootstrap/scss/utilities/display';
-@import 'node_modules/bootstrap/scss/utilities/embed';
-@import 'node_modules/bootstrap/scss/utilities/flex';
-@import 'node_modules/bootstrap/scss/utilities/float';
-@import 'node_modules/bootstrap/scss/utilities/position';
-@import 'node_modules/bootstrap/scss/utilities/screenreaders';
-@import 'node_modules/bootstrap/scss/utilities/sizing';
-@import 'node_modules/bootstrap/scss/utilities/spacing';
-@import 'node_modules/bootstrap/scss/utilities/text';
-@import 'node_modules/bootstrap/scss/utilities/visibility';
 @import 'gavant-ember-table/variables';
 
 .data-table {
@@ -63,15 +6,15 @@
     padding-left: 1.25rem;
     padding-right: 1.25rem;
 
-    @include media-breakpoint-up(lg) {
-        min-height: 30rem;
-    }
+    // @include media-breakpoint-up(lg) {
+    //     min-height: 30rem;
+    // }
 
-    &.static-table {
-        @include media-breakpoint-up(lg) {
-            min-height: initial;
-        }
-    }
+    // &.static-table {
+    //     @include media-breakpoint-up(lg) {
+    //         min-height: initial;
+    //     }
+    // }
 
     //since we can't detect how many columns to render in fastboot
     //make sure any overflow columns are hidden until the client re-renders
