+++ package.json
@@ -1,86 +1,89 @@
 {
-  "name": "@gavant/ember-table",
-  "version": "0.0.0",
-  "description": "Data tables built using ember-table",
-  "keywords": [
-    "ember-addon"
-  ],
-  "repository": "https://github.com/Gavant/gavant-ember-table",
-  "license": "MIT",
-  "author": "",
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
-  },
-  "scripts": {
-    "build": "ember build",
-    "lint:hbs": "ember-template-lint .",
-    "lint:js": "eslint .",
-    "start": "ember serve",
-    "test": "ember test",
-    "test:all": "ember try:each",
-    "prepublishOnly": "ember ts:precompile",
-    "postpublish": "ember ts:clean"
-  },
-  "dependencies": {
-    "ember-cli-babel": "^7.13.0",
-    "ember-cli-htmlbars": "^4.2.0",
-    "ember-cli-typescript": "^3.1.2"
-  },
-  "devDependencies": {
-    "@ember/optional-features": "^1.1.0",
-    "@glimmer/component": "^1.0.0",
-    "@types/ember": "^3.1.1",
-    "@types/ember-qunit": "^3.4.7",
-    "@types/ember__test-helpers": "^0.7.9",
-    "@types/qunit": "^2.9.0",
-    "@types/rsvp": "^4.0.3",
-    "babel-eslint": "^10.0.3",
-    "bootstrap": "^4.3.1",
-    "broccoli-asset-rev": "^3.0.0",
-    "ember-auto-import": "^1.5.3",
-    "ember-awesome-macros": "^5.0.1",
-    "ember-bootstrap": "^3.1.1",
-    "ember-cli": "~3.15.1",
-    "ember-cli-changelog": "^0.3.4",
-    "ember-cli-dependency-checker": "^3.2.0",
-    "ember-cli-eslint": "^5.1.0",
-    "ember-cli-github-pages": "^0.2.2",
-    "ember-cli-inject-live-reload": "^2.0.1",
-    "ember-cli-release": "^1.0.0-beta.2",
-    "ember-cli-sass": "^10.0.1",
-    "ember-cli-sri": "^2.1.1",
-    "ember-cli-template-lint": "^1.0.0-beta.3",
-    "ember-cli-typescript-blueprints": "^3.0.0",
-    "ember-cli-uglify": "^3.0.0",
-    "ember-disable-prototype-extensions": "^1.1.3",
-    "ember-export-application-global": "^2.0.1",
-    "ember-intl": "^4.2.2",
-    "ember-load-initializers": "^2.1.1",
-    "ember-maybe-import-regenerator": "^0.1.6",
-    "ember-qunit": "^4.6.0",
-    "ember-resize": "^0.3.4",
-    "ember-resolver": "^7.0.0",
-    "ember-source": "~3.15.0",
-    "ember-source-channel-url": "^2.0.1",
-    "ember-table": "^2.2.3",
-    "ember-try": "^1.4.0",
-    "eslint-plugin-ember": "^7.7.1",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "loader.js": "^4.7.0",
-    "prettier": "^1.19.1",
-    "qunit-dom": "^0.9.2",
-    "sass": "^1.24.4",
-    "typescript": "^3.7.4"
-  },
-  "engines": {
-    "node": "8.* || >= 10.*"
-  },
-  "ember": {
-    "edition": "octane"
-  },
-  "ember-addon": {
-    "configPath": "tests/dummy/config"
-  }
+    "name": "@gavant/ember-table",
+    "version": "0.0.0",
+    "description": "Data tables built using ember-table",
+    "keywords": [
+        "ember-addon"
+    ],
+    "repository": "https://github.com/Gavant/gavant-ember-table",
+    "license": "MIT",
+    "author": "",
+    "directories": {
+        "doc": "doc",
+        "test": "tests"
+    },
+    "scripts": {
+        "build": "ember build",
+        "lint:hbs": "ember-template-lint .",
+        "lint:js": "eslint .",
+        "start": "ember serve",
+        "test": "ember test",
+        "test:all": "ember try:each",
+        "prepublishOnly": "ember ts:precompile",
+        "postpublish": "ember ts:clean"
+    },
+    "dependencies": {
+        "ember-cli-babel": "^7.13.0",
+        "ember-cli-htmlbars": "^4.2.0",
+        "ember-cli-typescript": "^3.1.2",
+        "@gavant/ember-button-basic": "^0.0.9",
+        "ember-font-awesome": "^4.0.0-rc.4"
+    },
+    "devDependencies": {
+        "@ember/optional-features": "^1.1.0",
+        "@glimmer/component": "^1.0.0",
+        "@types/ember": "^3.1.1",
+        "@types/ember-qunit": "^3.4.7",
+        "@types/ember__test-helpers": "^0.7.9",
+        "@types/qunit": "^2.9.0",
+        "@types/rsvp": "^4.0.3",
+        "babel-eslint": "^10.0.3",
+        "broccoli-asset-rev": "^3.0.0",
+        "ember-auto-import": "^1.5.3",
+        "ember-awesome-macros": "^5.0.1",
+        "ember-cli": "~3.15.1",
+        "ember-cli-changelog": "^0.3.4",
+        "ember-cli-dependency-checker": "^3.2.0",
+        "ember-cli-eslint": "^5.1.0",
+        "ember-cli-github-pages": "^0.2.2",
+        "ember-cli-inject-live-reload": "^2.0.1",
+        "ember-cli-release": "^1.0.0-beta.2",
+        "ember-cli-sass": "^10.0.1",
+        "ember-cli-sri": "^2.1.1",
+        "ember-cli-string-helpers": "^4.0.6",
+        "ember-cli-template-lint": "^1.0.0-beta.3",
+        "ember-cli-typescript-blueprints": "^3.0.0",
+        "ember-cli-uglify": "^3.0.0",
+        "ember-disable-prototype-extensions": "^1.1.3",
+        "ember-export-application-global": "^2.0.1",
+        "ember-intl": "^4.2.2",
+        "ember-load-initializers": "^2.1.1",
+        "ember-maybe-import-regenerator": "^0.1.6",
+        "ember-qunit": "^4.6.0",
+        "ember-resize": "^0.3.4",
+        "ember-resolver": "^7.0.0",
+        "ember-responsive": "^3.0.5",
+        "ember-source": "~3.15.0",
+        "ember-source-channel-url": "^2.0.1",
+        "ember-table": "^2.2.3",
+        "ember-truth-helpers": "^2.1.0",
+        "ember-try": "^1.4.0",
+        "eslint-plugin-ember": "^7.7.1",
+        "eslint-plugin-node": "^10.0.0",
+        "eslint-plugin-prettier": "^3.1.2",
+        "loader.js": "^4.7.0",
+        "prettier": "^1.19.1",
+        "qunit-dom": "^0.9.2",
+        "sass": "^1.24.4",
+        "typescript": "^3.7.4"
+    },
+    "engines": {
+        "node": "8.* || >= 10.*"
+    },
+    "ember": {
+        "edition": "octane"
+    },
+    "ember-addon": {
+        "configPath": "tests/dummy/config"
+    }
 }
