+++ tests/dummy/app/config/environment.d.ts
@@ -8,9 +8,9 @@ export default config;
  * since different ember addons can materialize new entries.
  */
 declare const config: {
-  environment: any;
-  modulePrefix: string;
-  podModulePrefix: string;
-  locationType: string;
-  rootURL: string;
+    environment: any;
+    modulePrefix: string;
+    podModulePrefix: string;
+    locationType: string;
+    rootURL: string;
 };
