+++ tests/dummy/app/styles/app.scss
@@ -1,0 +1,1 @@
+@import 'gavant-ember-table';
