+++ types/ember-responsive/index.d.ts
@@ -1,0 +1,8 @@
+declare module 'ember-responsive' {
+    export default class Media {
+        isMobile: boolean;
+        isTablet: boolean;
+        isDesktop: boolean;
+        isJumbo: boolean;
+    }
+}
