+++ types/ember-table/index.d.ts
@@ -1,0 +1,13 @@
+declare module 'ember-table' {
+    interface TableColumn {
+        id?: string;
+        valuePath: string;
+        name: string;
+        isFixedLeft: boolean;
+        width: number;
+        staticWidth: number;
+        minWidth?: number;
+        maxWidth?: number;
+        textAlign?: string;
+    }
+}
